LNDFS is back
Only added that each thread gets a successor of the initial state (1st thread the 1st successor, 2nd the 2nd.....). If all successors are gone, the successors of the initial state's last successor are used. (since first thread would got the first successor first and is probably the furthest in the graph, last successor is probably the furthest behind)